Refactor: Reorganise reducers for readability. Add Contract reducers. It works.
