+++ src/components/SmartContracts/ContractsMain/ContractsMain.tsx
@@ -1,7 +1,23 @@
+import { useSelector, useDispatch } from "react-redux"
+import { getAllContracts } from "../../../store/actions"
+
+import ContractsDataBlock from "../ContractsDataBlock/ContractsDataBlock"
+
+import { State } from "../../../utility/types"
+
 import "./ContractsMain.css"
 
 const ContractsMain = () => {
-  return <main className="contracts-main">SMART CONTS HERE</main>
+  const activeChannel = useSelector((state: State) => state.basic.activeChannel)
+  const activeChannelHash = activeChannel.channel_genesis_hash
+  const dispatch = useDispatch()
+  dispatch(getAllContracts(activeChannelHash.toString()))
+
+  return (
+    <main className="contracts-main">
+      <ContractsDataBlock />
+    </main>
+  )
 }
 
 export default ContractsMain
