+++ src/store/actions/actionTypes.tsx
@@ -8,3 +8,4 @@ export const BLOCK_ACTIVITY_DATA = "BLOCK_ACTIVITY_DATA"
 export const ADD_NEW_BLOCK = "ADD_NEW_BLOCK"
 export const TXN_ACTIVITY_DATA = "TXN_ACTIVITY_DATA"
 export const ADD_NEW_TXNS = "ADD_NEW_TXNS"
+export const GET_ALL_CONTRACTS = "GET_ALL_CONTRACTS"
