+++ src/store/actions/index.ts
@@ -1,0 +1,11 @@
+export { addNewBlock, setBlockActivityData } from "./blockActions"
+export { addNewTxns, setTxnActivityData } from "./txnActions"
+export { getAllContracts } from "./contractActions"
+export {
+  setChannelStats,
+  getAvailableChannels,
+  setChannelInfo,
+  setServerStatus,
+  setActiveChannel,
+  setChannelHash,
+} from "./basicActions"
